# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8444
  description: Generated server url
paths:
  /api/inventory/items/{id}/update:
    put:
      tags:
      - items-rest-controller
      summary: Update an item
      operationId: itemUpdating
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: imageField
        in: query
        required: true
        schema:
          type: string
          format: binary
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        "404":
          description: No item updated
        "200":
          description: Update an item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
  /api/inventory/sizes:
    get:
      tags:
      - items-rest-controller
      summary: Get sizes
      operationId: getSizes
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Get sizes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Size'
        "404":
          description: No size founded
    post:
      tags:
      - items-rest-controller
      summary: Post a new size
      operationId: addSize
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Size'
        required: true
      responses:
        "200":
          description: Post a new size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Size'
        "404":
          description: No size posted
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Size'
  /api/inventory/items:
    post:
      tags:
      - items-rest-controller
      summary: Post a new item
      operationId: addItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        "200":
          description: Post a new item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: No item posted
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Item'
  /api/inventory/items/{id}/image:
    post:
      tags:
      - items-rest-controller
      summary: Post an item image
      operationId: addItemImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              required:
              - itemImage
              type: object
              properties:
                itemImage:
                  type: string
                  format: binary
      responses:
        "200":
          description: Post an item image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: No item image posted
  /api/inventory/items/shoes/stock:
    post:
      tags:
      - items-rest-controller
      summary: Post a new stock of type shoes
      operationId: addShoeStock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shoe'
        required: true
      responses:
        "200":
          description: Post a new stock of type shoes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shoe'
        "404":
          description: No shoes posted
  /api/inventory/items/clothes/stock:
    post:
      tags:
      - items-rest-controller
      summary: Post a new stock of type clothes
      operationId: addClothesStock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Clothes'
        required: true
      responses:
        "200":
          description: Post a new stock of type clothes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clothes'
        "404":
          description: No clothes posted
  /redis/items/{id}:
    get:
      tags:
      - items-rest-controller
      operationId: getItemByIdReddis
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Item'
  /api/inventory/stocks:
    get:
      tags:
      - items-rest-controller
      summary: Get stocks
      operationId: getAllStocks
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "404":
          description: No stock founded
        "200":
          description: Post a new stocks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
  /api/inventory/stocks/item/{id}:
    get:
      tags:
      - items-rest-controller
      summary: Get an item stocks
      operationId: getStockItem
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No item stocks founded
        "200":
          description: Get an item stocks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
  /api/inventory/shoes/stock:
    get:
      tags:
      - items-rest-controller
      summary: Get stock of type shoes
      operationId: getShoesStock
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "404":
          description: No shoes founded
        "200":
          description: Post a new stock of type shoes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shoe'
  /api/inventory/items/{id}:
    get:
      tags:
      - items-rest-controller
      summary: Get an item
      operationId: getItemById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Get an item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: No item founded
  /api/inventory/items/{id1}/stocks/{id2}/sizes/{id3}:
    get:
      tags:
      - items-rest-controller
      summary: "Post an item, stock and size"
      operationId: addSizeAndStockToItem
      parameters:
      - name: id1
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: id2
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: id3
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: "Post an item, stock and size"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: No item founded
  /api/inventory/clothes/stock:
    get:
      tags:
      - items-rest-controller
      summary: Get stock of type shoes
      operationId: getClothesStock
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "404":
          description: No shoes founded
        "200":
          description: Get stock of type shoes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shoe'
  /api/inventory/stocks/{id}:
    delete:
      tags:
      - items-rest-controller
      summary: Delete a stock constraints
      operationId: deleteRelationshipItemStock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No stock deleted
        "200":
          description: Delete a stock constraints
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
  /api/inventory/stocks/{id}/delete:
    delete:
      tags:
      - items-rest-controller
      summary: Delete a stock
      operationId: deleteItemStock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Delete a stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        "404":
          description: No stock deleted
  /api/inventory/items/{id}/delete:
    delete:
      tags:
      - items-rest-controller
      summary: Delete an item
      operationId: deleteItem
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Delete an item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: No item deleted
components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        gender:
          type: string
        type:
          type: string
        imageFile:
          type: object
          properties:
            binaryStream:
              type: object
        stock:
          $ref: '#/components/schemas/StockObject'
        stocks:
          type: array
          items:
            $ref: '#/components/schemas/StockObject'
    Size:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        label:
          type: string
    StockObject:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        stock:
          type: integer
          format: int32
        size:
          $ref: '#/components/schemas/Size'
    Shoe:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        stock:
          type: integer
          format: int32
        size:
          $ref: '#/components/schemas/Size'
    Clothes:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        stock:
          type: integer
          format: int32
        size:
          $ref: '#/components/schemas/Size'
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    Stock:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        stock:
          type: integer
          format: int32
        size:
          $ref: '#/components/schemas/Size'